rename js snake_case to CamelCase
