+++ public/src/js/app.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import $ from 'jquery';
-import SlackApp from './components/slack-app.js';
+import SlackApp from './components/SlackApp';
 
 $(() => {
   React.render(<SlackApp />, $('#app').get(0));
