+++ public/src/js/components/SlackChannel.js
