+++ public/src/js/components/SlackMessage.js
